requirements.txt
openai==0.28.0
pandas==2.2.3
pgvector==0.3.6
python-dotenv==1.0.1
schedule==1.2.2
slack_sdk==3.33.3
SQLAlchemy==2.0.36
Telethon==1.38.1
tiktoken==0.8.0
tqdm==4.67.0
